Added changes.

+++ prework/src/pages/EditCreator.jsx
@@ -1,24 +1,25 @@
 import React from 'react';
 import { Link, useParams } from "react-router-dom";
 import { useEffect, useState } from 'react';
-import ReadPost from './ViewCreator.jsx'
+import ViewCreator from './ViewCreator.jsx'
 import { supabase } from '../components/Client.jsx'
 
-const EditPost = () => {
+const EditCreator = () => {
     const [name, setName] = useState("");
     const [speed, setSpeed] = useState(0);
     const [color, setColor] = useState("plum");
     const [error, setError] = useState(null);
-    const [post, setPost] = useState({id: 0, name: "", title: "", date: ""});
+    const [creator, setCreator] = useState({id: 0, name: "", url: "", image_url: "", description: ""});
     const [id, setId] = useState(0);
 
-    const updatePost = async (event) => {
+    const updateCreator = async (event) => {
         event.preventDefault();
 
         const { data, error } = await supabase 
-            .from("Posts")
-            .update({ title: post.name, created_at: post.date })
-            .eq('id', post.id)
+            .from("Creators")
+            .update({ name: creator.name, url: creator.url, 
+            image_url: creator.image_url, description: creator.description })
+            .eq('id', creator.id)
 
         if (error) {
             setError("Need to enter information.");
@@ -31,8 +32,8 @@ const EditPost = () => {
     }
 
     const handleChange = (event) => {
-        const {name, value} = event.target;
-        setPost((prev) => {
+        const { name, value } = event.target;
+        setCreator((prev) => {
             return {
                 ...prev,
                 [name]:value,
@@ -40,13 +41,13 @@ const EditPost = () => {
         })
     }
 
-    const deletePost = async (event) => {
+    const deleteCreator = async (event) => {
         event.preventDefault();
       
         await supabase
-          .from('Posts')
+          .from('Creators')
           .delete()
-          .eq('id', post.id); 
+          .eq('id', creator.id); 
       
         window.location = "http://localhost:5173/";
     }
@@ -56,40 +57,30 @@ const EditPost = () => {
             <div>
                 <form className='forms'>
                     <label htmlFor='id'> Post Id: </label>
-                    <input 
-                    type='text'
-                    id='id'
-                    name='id'
-                    value={post.id}
-                    onChange={handleChange} />
-                    <br/>
-                    <br/>
+                    <input type='text' id='id' name='id' value={creator.id}
+                    onChange={handleChange} /> <br/> <br/>
 
                     <label htmlFor='name'> Name: </label>
-                    <input 
-                    type='text'
-                    id='title'
-                    name='name'
-                    value={post.name}
-                    onChange={handleChange} />
-                    <br/>
-                    <br/>
+                    <input type='text' id='name' name='name' value={creator.name} 
+                    onChange={handleChange} /> <br/> <br/>
 
-                    <label htmlFor='date'> Date: </label>
-                    <input 
-                    type='description'
-                    id='description'
-                    name='date'
-                    value={post.date}
-                    onChange={handleChange} />
-                    <br/>
-                    <br/>
+                    <label htmlFor='date'> Url: </label>
+                    <input type='text' id='description' name='url' value={creator.url}
+                    onChange={handleChange} /> <br/> <br/>
+
+                    <label htmlFor='date'> ImageUrl: </label>
+                    <input type='text' id='description' name='image_url' value={creator.image_url}
+                    onChange={handleChange} /> <br/> <br/>
+
+                    <label htmlFor='date'> Description: </label>
+                    <input type='description' id='description' name='description' value={creator.descrition}
+                    onChange={handleChange} /> <br/> <br/>
                     
                     <label htmlFor="red">Red</label><br />
                     <input type="radio" id="red" name="color" value="red" onChange={handleChange} /><br />
 
-                    <input className="inputs" type="submit" value="Edit a Creator" onClick={updatePost}/>
-                    <input className="inputs" type="submit" value="Delete a Creator" onClick={deletePost}/>
+                    <input className="inputs" type="submit" value="Edit a Creator" onClick={updateCreator}/>
+                    <input className="inputs" type="submit" value="Delete a Creator" onClick={deleteCreator}/>
                 </form>
             </div>     
         </>     
@@ -97,4 +88,4 @@ const EditPost = () => {
 
 };
 
-export default EditPost;
+export default EditCreator;
